+++ packages/version/src/version-command.ts
@@ -29,13 +29,16 @@ import semver from 'semver';
 import c from 'tinyrainbow';
 import zeptomatch from 'zeptomatch';
 
+import { COMMENT_FILTER_KEYWORDS_CSV, COMMENT_ISSUE, COMMENT_PULL_REQUEST } from './constant.js';
 import { applyBuildMetadata } from './conventional-commits/apply-build-metadata.js';
 import { getCommitsSinceLastRelease } from './conventional-commits/get-commits-since-last-release.js';
 import { recommendVersion } from './conventional-commits/recommend-version.js';
 import { updateChangelog } from './conventional-commits/update-changelog.js';
-import type { GitCreateReleaseClientOutput, ReleaseNote, RemoteCommit } from './interfaces.js';
+import type { OctokitClientOutput, ReleaseNote, RemoteCommit } from './interfaces.js';
+import { commentResolvedItems } from './lib/comment-resolved-items.js';
 import { createRelease, createReleaseClient } from './lib/create-release.js';
 import { getCurrentBranch } from './lib/get-current-branch.js';
+import { getPreviousTag, type PreviousTag } from './lib/get-tag.js';
 import { gitAdd } from './lib/git-add.js';
 import { gitCommit } from './lib/git-commit.js';
 import { gitPush, gitPushSingleTag } from './lib/git-push.js';
@@ -71,9 +74,10 @@ export class VersionCommand extends Command<VersionCommandOption> {
   commitAndTag = true;
   pushToRemote = true;
   hasRootedLeaf = false;
-  releaseClient?: GitCreateReleaseClientOutput;
+  releaseClient?: OctokitClientOutput;
   releaseNotes: ReleaseNote[] = [];
   gitOpts: any;
+  lastTagCommit?: PreviousTag;
   runPackageLifecycle: any;
   runRootLifecycle!: (stage: string) => Promise<void> | void;
   savePrefix = '';
@@ -292,6 +296,9 @@ export class VersionCommand extends Command<VersionCommandOption> {
       );
     }
 
+    // keep last tag oldest commit details as reference
+    this.lastTagCommit = getPreviousTag(this.execOpts, isIndependent);
+
     // fetch all commits from remote server of the last release when user wants to include client login associated to each commits
     const remoteClient = this.options.createRelease || this.options.remoteClient;
     if (this.options.conventionalCommits && this.changelogIncludeCommitsClientLogin) {
@@ -427,6 +434,11 @@ export class VersionCommand extends Command<VersionCommandOption> {
 
     if (!this.composed) {
       this.logger.success('version', 'finished');
+    }
+
+    // comment on remote issues/PRs
+    if (this.options.commentIssues || this.options.commentPullRequests) {
+      await this.commentOnRemote();
     }
 
     return {
@@ -914,6 +926,47 @@ export class VersionCommand extends Command<VersionCommandOption> {
     // or push all tags by using followTags
     this.logger.info('git', 'Pushing tags...');
     return gitPush(this.gitRemote, this.currentBranch, this.execOpts, this.options.dryRun);
+  }
+
+  /** Comment on resolved issues and/or merged PRs */
+  async commentOnRemote() {
+    const clientType = this.options.createRelease || this.options.remoteClient;
+    if (clientType === 'gitlab') {
+      this.logger.warn('comments', 'GitLab is not currently supported to comment on issues/PRs.');
+      return Promise.resolve();
+    } else if (clientType === 'github') {
+      const client = this.releaseClient || (await createReleaseClient(clientType));
+      const {
+        dryRun = false,
+        gitRemote = 'origin',
+        tagVersionPrefix = 'v',
+        commentIssues,
+        commentPullRequests,
+        commentFilterKeywords: keywordsCSV = COMMENT_FILTER_KEYWORDS_CSV,
+      } = this.options;
+
+      const logPrefix = dryRun ? c.bgMagenta('[dry-run] ') : '';
+      this.logger.info('comments', `${logPrefix}[start] Comments on remote client...`);
+
+      await commentResolvedItems({
+        client,
+        commentFilterKeywords: keywordsCSV.split(','),
+        dryRun,
+        gitRemote,
+        execOpts: this.execOpts,
+        prevTagDate: this.lastTagCommit?.date || '',
+        logger: this.logger,
+        tag: `${tagVersionPrefix}${this.globalVersion}`,
+        version: this.globalVersion,
+        templates: {
+          // use custom template when provided by the user or use default template when enabled via boolean (empty when false or undefined)
+          issue: (commentIssues === true ? COMMENT_ISSUE : commentIssues) || '',
+          pullRequest: (commentPullRequests === true ? COMMENT_PULL_REQUEST : commentPullRequests) || '',
+        },
+      });
+
+      this.logger.info('comments', '[end] Comments on remote client...');
+    }
   }
 
   setGlobalVersionFloor() {