+++ CHANGELOG.md
@@ -4,6 +4,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.1](https://github.com/lerna-lite/lerna-lite/compare/v4.10.0...v4.10.1) (2025-12-12)
+
+### Bug Fixes
+
+* delay issue comments between each creation calls & add permissions ([6079ea1](https://github.com/lerna-lite/lerna-lite/commit/6079ea12fea3d320772a907a08f7197d21b21194)) - by @ghiscoding
+* move Comments on Issues/PRs from lerna publish to version ([#1208](https://github.com/lerna-lite/lerna-lite/issues/1208)) ([c2d8e7a](https://github.com/lerna-lite/lerna-lite/commit/c2d8e7ae71ee58fccd6ff5213c309b7c2b954569)) - by @ghiscoding
+* **version:** use last tag date for comment issues/PRs querying ([#1209](https://github.com/lerna-lite/lerna-lite/issues/1209)) ([92627de](https://github.com/lerna-lite/lerna-lite/commit/92627dec57daf0154b85ac71e44ef8cce39a563f)) - by @ghiscoding
+
+## [4.10.0](https://github.com/lerna-lite/lerna-lite/compare/v4.9.4...v4.10.0) (2025-12-12)
+
+### Features
+
+* **publish:** Comment on issues/PRs resolved by release, fix [#1198](https://github.com/lerna-lite/lerna-lite/issues/1198) ([#1200](https://github.com/lerna-lite/lerna-lite/issues/1200)) ([5b0cef9](https://github.com/lerna-lite/lerna-lite/commit/5b0cef9f86cf3c798927dcd2cdb73444e9d00b80)) - by @ghiscoding
+* verify OIDC in both dry-run and normal publish mode ([#1195](https://github.com/lerna-lite/lerna-lite/issues/1195)) ([4286a59](https://github.com/lerna-lite/lerna-lite/commit/4286a59b7e18f6006022e204d23c50d5e79bb8e5)) - by @jcpoirier20
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies ([#1201](https://github.com/lerna-lite/lerna-lite/issues/1201)) ([0010e25](https://github.com/lerna-lite/lerna-lite/commit/0010e253f183ff4e8e226a48526069184f45ad44)) - by @renovate[bot]
+* **publish:** shouldn't Comment on closed issues older than release ([#1202](https://github.com/lerna-lite/lerna-lite/issues/1202)) ([199bfed](https://github.com/lerna-lite/lerna-lite/commit/199bfedbded089a6edb357047d56f46dda2547c0)) - by @ghiscoding
+
 ## [4.9.4](https://github.com/lerna-lite/lerna-lite/compare/v4.9.3...v4.9.4) (2025-11-27)
 
 ### Bug Fixes