+++ packages/list/CHANGELOG.md
@@ -4,6 +4,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.1](https://github.com/lerna-lite/lerna-lite/compare/v4.10.0...v4.10.1) (2025-12-12)
+
+**Note:** Version bump only for package @lerna-lite/list
+
+## [4.10.0](https://github.com/lerna-lite/lerna-lite/compare/v4.9.4...v4.10.0) (2025-12-12)
+
+**Note:** Version bump only for package @lerna-lite/list
+
 ## [4.9.4](https://github.com/lerna-lite/lerna-lite/compare/v4.9.3...v4.9.4) (2025-11-27)
 
 **Note:** Version bump only for package @lerna-lite/list